+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `withPlayerButtonOnImage` props and update style on
+## [1.20.0] - 2019-03-08
+
+Feature:
+- Add `withPlayerButtonOnImage` props and update style on
   `SimpleCard` component.
 
 ## [1.19.0] - 2019-03-05